pyproject.toml
[tool.poetry]
name = "podgpt"
version = "0.1.0"
description = ""
authors = ["podrob <rob.mcbryde@pod-point.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
python-dotenv = "^1.0.0"
langchain = "^0.0.302"
tqdm = "^4.66.1"
faiss-cpu = "^1.7.4"
streamlit = "^1.28.2"
streamlit-chat = "^0.1.1"
pypdf = "^3.17.1"
docx2txt = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
